bug fixes
this takes 42 seconds to find the maximum resnik similarity across all synsets of two words using the brown corpus (or around 4 seconds for path similarity), so I'm not sure how much more development will go into using nltk's wordnet architecture here. The goal for this project was to find a reasonably time-efficient means of judging the similarity of two words.